+++ CHANGELOG.md
@@ -4,6 +4,23 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+## [0.0.2] - 2023-01-16
+
+### Added
+- toggle button for dark/light theme
+- header and header style
+- footer and footer style
+- general styles to index.css
+- main view folder, with Main.js and main.css
+- header view folder, with Header.js and header.css
+- footer view folder, with Footer.js and footer.css
+- views folder, which will contain the different parts of the app.
+- colors.css. This file contains colors for light and dark theme.
+- src/styles folder, which will contain general style files
+
+### Changed
+- Moved index.css into src/styles
+
 ## [0.0.1] - 2023-01-15
 
 ### Added